pyproject.toml
[project]
name = "realitylens-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.70.0",
    "google-generativeai>=0.8.6",
    "keyboard>=0.13.5",
    "pillow>=12.2.0",
    "pyautogui>=0.9.54",
    "pyinstaller>=6.19.0",
    "pyqt6>=6.11.0",
    "pyscreeze>=1.0.1",
    "python-dotenv>=1.2.2",
    "uv>=0.11.3",
]
